Pull daily (#6343)
* daily pull to archival job

* new bucket system

* pattern mine bucket fixes

* new bucket system

* added merge channel on beam and code cleanup

* added group info when pulling users and bugs fixes

* minor fix

* minor fix

* fixes and task manager changes

* flag protect, new jobs and minor fixes

* server fixes

* flag protect jobs for using new buckets

* pull events job flag protect incomplete

* flag protect pull events and added sorting on groups

* minor fixes

* proper task names

* minor fixes

* fixing path analysis script

* fixes

* soft pull to check 2 days for data

* dockerfile changes to solve beam error

* minor fixes

* beam issue fixed

* path analysis - adding projectids

* explain v2 minor fix

* pulling group_id instead of group_user_id

* adding groups support for path analysis

* added documentation for functions

* Revert "pulling group_id instead of group_user_id"

This reverts commit af46d2333757e4204b655201ebdbe02ce032d237.

* adding group userid

* adding group event support

* comments and logging

* code quality

* reverse merge fixes

* fix warnings

* minor fixes and code quality

* delta test fix and code quality

* new flow local fix

* local fix

+++ backend/src/factors/scripts/adhoc/run_delete_dangling_sessions/run_delete_dangling_sessions.go
@@ -181,7 +181,7 @@ func main() {
 		unusedSessionIDs, errCode = getUnusedSessionIDsFromArchiveFile(cloudStorage, *archiveEventsFilepath, *archiveEventsFilename)
 	} else {
 		logCtx.Info("Reading events from DB.")
-		unusedSessionIDs, errCode = store.GetStore().GetUnusedSessionIDsForJob(*projectID, *startTimestamp, *endTimestamp)
+		unusedSessionIDs, errCode = store.GetStore().GetUnusedSessionIDsForJob(int64(*projectID), *startTimestamp, *endTimestamp)
 	}
 
 	if errCode == http.StatusInternalServerError {
@@ -199,14 +199,14 @@ func main() {
 		return
 	}
 
-	sessionEventName, errCode := store.GetStore().GetSessionEventName(*projectID)
+	sessionEventName, errCode := store.GetStore().GetSessionEventName(int64(*projectID))
 	if errCode != http.StatusFound {
 		logCtx.Error("Failed to get session event_name.")
 		return
 	}
 
 	logCtx.Info("Started deleting dangling sessions.")
-	errCode = store.GetStore().DeleteEventsByIDsInBatchForJob(*projectID, sessionEventName.ID, unusedSessionIDs, 1000)
+	errCode = store.GetStore().DeleteEventsByIDsInBatchForJob(int64(*projectID), sessionEventName.ID, unusedSessionIDs, 1000)
 	if errCode == http.StatusInternalServerError {
 		log.Error("Failed to delete sessions in batch")
 		return
